Switched number of shares to BigDecimal. Removed stockCurrentPrice from delete.html.

+++ src/main/java/com/inprogress/portfolio/commands/StockCommand.java
@@ -22,7 +22,7 @@ public class StockCommand {
     @Size(min = 1, max = 10)
     private String stockSymbol;
 
-    private Integer numberOfShares;
+    private BigDecimal numberOfShares;
 
     private BigDecimal stockPurchasePrice;
